+++ packages/terra-button/src/Button.jsx
@@ -80,7 +80,10 @@ const Button = ({
   ...customProps
   }) => {
   const attributes = Object.assign({}, customProps);
-  const buttonText = text ? <span className={styles.text}>{text}</span> : null;
+
+  const iconWithTextStyle = isReversed ? cx(['text-with-icon-reverse']) : cx(['text-with-icon']);
+  const textStyle = icon ? iconWithTextStyle : null;
+  const buttonText = text ? <span className={textStyle} >{text}</span> : null;
 
   attributes.className = cx([
     'button',